rename package to data-provider
