+++ apps/server/src/config/server.ts
@@ -1,6 +1,6 @@
 import 'dotenv/config';
 
-import { AppConfig } from 'provider';
+import { AppConfig } from 'data-provider';
 
 export const getServerConfig = () => {
 	return {
