+++ package.json
@@ -6,7 +6,8 @@
 		"\n=================  Build  ===============": "",
 		"build:all": "pnpm -r build:all",
 		"\n============  Dev environment  ==========": "",
-		"start:dev": "pnpm --filter data-model db:migrate:dev && pnpm --parallel start:dev",
+		"with-env": "dotenv -e ./.env --",
+		"start:dev": "pnpm with-env pnpm --filter data-model db:migrate:dev && pnpm with-env pnpm --parallel start:dev",
 		"\n====  Prod environment (compiled)  ======": "",
 		"start:prod": "pnpm --filter data-model db:migrate:workspace-prod && pnpm --filter server start:worskpace-prod"
 	},
@@ -14,6 +15,7 @@
 	"license": "ISC",
 	"devDependencies": {
 		"@types/node": "^20.11.16",
+		"dotenv-cli": "^7.4.1",
 		"prettier": "^3.2.5",
 		"prettier-plugin-organize-imports": "^3.2.4",
 		"ts-node": "^10.9.2",
