+++ packages/data-provider/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "provider",
+	"name": "data-provider",
 	"version": "1.0.0",
 	"description": "",
 	"main": "dist/index.js",
