+++ packages/data-provider/src/controllers/dictionaryController.ts
